LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.font.weight-expected.txt

2d.text.font.weight
Actual output:

FAIL Canvas test: 2d.text.font.weight assert_equals: ctx.font === 'italic 300 12px serif' (got italic 12px serif[string], expected italic 300 12px serif[string]) expected "italic 300 12px serif" but got "italic 12px serif"
